Fix sign-compare in Transaction::processRequestBody
Use uint64_t for reqbodyNoFilesLength when comparing against
SecRequestBodyNoFilesLimit.

+++ src/transaction.cc
@@ -770,7 +770,7 @@ int Transaction::processRequestBody() {
     if (m_requestBodyType == MultiPartRequestBody) {
 #endif
         std::string error;
-        int reqbodyNoFilesLength = 0;
+        uint64_t reqbodyNoFilesLength = 0;
         if (a != NULL) {
             Multipart m(*a, this);
             if (m.init(&error) == true) {
